src/requirements.txt
argon2-cffi>=21.*
Flask>=2.*
